backend/requirements.txt
fastapi
uvicorn[standard]==0.23.0
python-multipart
smolagents==0.1.0
litellm
pandas
numpy==1.24.0
matplotlib==3.7.0
seaborn==0.12.0
scikit-learn==1.3.0
#pycaret==3.0.0
torch
torchvision
#pyradiomics
#SimpleITK==2.2.0
nibabel==5.0.0
PyYAML==6.0
python-dotenv
openai
anthropic==0.18.0